requirements.txt
pandas==1.3.4
Pillow==9.2.0
streamlit==1.9.2
streamlit_timeline==0.0.2
